weather_frontend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
requests==2.32.3
aiohttp==3.10.5

